server/src/requirements.txt
fastapi==0.110.0
openai==1.14.2
PyMuPDF==1.23.26
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn==0.27.1
